Fix map roundtrip issues

+++ src/OpenSage.Game/Data/Map/BuildListItem2.cs
@@ -8,7 +8,7 @@ namespace OpenSage.Data.Map
     /// </summary>
     public sealed class BuildListItem2
     {
-        public uint Unknown1 { get; private set; }
+        public ushort Unknown1 { get; private set; }
         public string BuildingName { get; private set; }
         public byte[] Unknown2 { get; private set; }
